change coordinate type in wasm example to float

+++ example/wasm/src/model.hpp
@@ -4,7 +4,7 @@
 #include <boost/geometry.hpp>
 
 namespace model {
-    using point = boost::geometry::model::point<double, 2, boost::geometry::cs::cartesian>;
+    using point = boost::geometry::model::point<float, 2, boost::geometry::cs::cartesian>;
     using trajectory = boost::geometry::model::linestring<point>;
     using vector = point;
 } // namespace model
